Split handshake tests into 2 files to prevent state machine conflicts
